Add slugify to the project and implement it in the tourSchema
This commit adds the slugify package to the project and uses it in the tourSchema to generate a slug for each tour when it is saved. The pre middleware function is used to generate the slug based on the tour name and then assign it to the slug field in the tourSchema. The lower option is passed to slugify to convert the slug to lowercase.