Add dynamic tour.pug template and route for individual tours
Description: Created a dynamic tour.pug template and corresponding route in the router to allow users to view individual tours by slug. The route uses the slug parameter to query the database for the corresponding tour data and render it in the tour.pug template. This allows users to view more detailed information about a specific tour. The route is defined as follows: router.route("/tour/:slug").get(viewController.getTour);