Remove unused files from the public folder inside the JS directory
This commit removes any unnecessary files from the public folder that was located within the JS directory. These files were not being used and were taking up unnecessary space, so they were safely removed from the project. This will improve the overall efficiency of the project by reducing the size of the repository and making it easier to manage.