Add login functionality with login.pug and login.js
This commit introduces a new login.pug file to the views folder, which contains the HTML markup for the login page. In addition, a new login.js file has been created in the public directory, which contains the client-side JavaScript code for handling login functionality. The login.js file communicates with the server to authenticate user credentials and provides feedback to the user if the login attempt was successful or not. This new functionality will allow users to securely log in to the website and access protected content.