Update Content Security Policy to allow connections to localhost:8080"  Description: Added the following lines of code to the server-side file:
This updates the Content Security Policy (CSP) to allow connections to http://localhost:8080. The connect-src directive is set to allow connections to the same origin ('self') and to http://localhost:8080. This change was necessary to resolve a Refused to connect error when trying to connect to http://localhost:8080/api/v1/users/login.