Create account.pug and add protected route for account page
In this commit, I created a new Pug file called account.pug in the views folder. Additionally, I added a new protected route to the router for the account page at /me, which requires authentication before rendering the page.

To render the account page, I also added an exports.getAccount function that sends a 200 status and renders the account.pug template with a title of "Your Account".