Add functionality to update user password
his commit adds the functionality to update user password. A listener is added to the password form in the index.js file. The updateSettings function is called with the password values entered by the user. In the updateSettings.js file, the fetch method is used to send the updated password data to the appropriate endpoint. The showAlert function is called to display a success or error message based on the response from the server. Additionally, a photo from the users folder in the public folder has been deleted.

Changes:

Created event listener for password form submission
Called updateSettings function with password data
Updated fetch method to send data to correct endpoint
Added showAlert function to display success/error message
Deleted a photo from the users folder inside public folder