Add form submit event listener and updateSettings function call for user data
Added a submit event listener to the user data form to handle updates to the user's name, email, and profile photo. The form data is collected using the FormData API and passed to the updateSettings function along with the string "data" to indicate that this is an update to the user's data. The updateSettings function then sends a PATCH request to the server to update the user's information.