Add image upload functionality to update tour route
Add image upload functionality to update tour route

In order to allow admins and lead guides to update tours with images, this commit adds the uploadTourImages and resizeTourImages middleware functions to the PATCH /api/v1/tours/:id route. This allows image files to be uploaded and resized before being saved to disk, and their filenames to be added to the req.body object. Finally, the updateTour controller function is called to save the updated tour information to the database. The route is protected with authentication and authorization middleware to ensure only authorized users can access it.