Add Signup Page and Functionality
Create signup.pug file in the views folder, which contains the signup form with required input fields and a submit button. Also, created signup.js file in the public/js folder, which contains the signup function that fetches data from the server and shows success or error message alerts based on the response.

In addition, added a functionality and render route for the signup page in the viewsRoutes.js file inside the routes folder. This new route renders the signup view and sends the necessary data to the viewController to display the signup form.