Create stripe.js page and integrate with index.js
In stripe.js, added the bookTour function that uses fetch API to get the checkout session from the API and redirects the user to the checkout URL. If an error occurs, it shows an alert message.

In index.js, imported the bookTour function from stripe.js and added a click event listener to the book button. When clicked, it changes the text to "Processing...", gets the tour ID from the dataset attribute of the button, and calls the bookTour function with the tour ID.

Also imported showAlert function from alert.js in stripe.js.