Added Booking model schema in bookingModel.js file in the models folder
This commit adds the Mongoose schema for Booking model to the bookingModel.js file. It defines the schema with fields such as tour, user, price, createdAt, and paid. It also defines a pre-query middleware to populate the user and tour fields. Finally, it creates the Booking model using the schema and exports it for use in other parts of the application.